+++ src/screens/secrets/SecretsMenuScreen.tsx
@@ -1,15 +1,17 @@
 import { StyleSheet, View } from 'react-native';
 
+import { Icons } from '../../assets/icons';
 import {
   DashboardAction,
   SecretsMenuScreenProps,
 } from '../../navigation/types';
 import theme from '../../theme';
 
-import Menu from '../../components/Menu';
+import EntryList from '../../components/EntryList';
 
 export const dashboardEntry: DashboardAction = {
   label: 'Secrets',
+  icon: Icons.secrets,
   navigate: nav => nav.navigate('SecretsMenu'),
 };
 
@@ -19,14 +21,17 @@ export default function SecretsMenuScreen({
   const entries = [
     {
       label: 'Change PIN',
+      icon: Icons.pin,
       onPress: () => navigation.navigate('ChangeSecret', { secretType: 'pin' }),
     },
     {
       label: 'Change PUK',
+      icon: Icons.puk,
       onPress: () => navigation.navigate('ChangeSecret', { secretType: 'puk' }),
     },
     {
       label: 'Change Pairing Secret',
+      icon: Icons.pairingSecret,
       onPress: () =>
         navigation.navigate('ChangeSecret', { secretType: 'pairing' }),
     },
@@ -34,7 +39,7 @@ export default function SecretsMenuScreen({
 
   return (
     <View style={styles.container}>
-      <Menu entries={entries} />
+      <EntryList entries={entries} />
     </View>
   );
 }